requirements.txt
kafka-python==2.0.2
redis==5.0.1
dataclasses-json==0.6.3